fix: move Solution from Problem1.java to Problem2.java
